Bug #41: Include guards became standard-compliant and unique.
Some underscores were replaced to make affected identifiers standard-compliant.

Include guards prevent the multiple inclusion of header files only correctly
if the chosen identifiers are different from others with such a functionality.
The probability for name clashes was reduced by the addition of the project name
as a prefix and a kind of universally unique identifier as a suffix.

Signed-off-by: Markus Elfring <elfring@users.sourceforge.net>