CS_OPT_CHECKINSN improved.
(1) Callback needs more parameters to handle cases similar to SKIPDATA but with more flexibility to be able to skip a data whether it is a decodable instruction or not.

(2) Callback returns false to stop disassembly and discard the current instruction which was passed to the callback from the result.

(3) If callback is called with insn->id == 0, the user callback must fill:
	- insn->address (with offset passed as parameter),
	- insn->size (it should be equal or less  than code_size passed as parameter),
	- insn->bytes (it should be copied from code passed as parameter and length must be equal or less than code_size),
	- insn->mnemonic (at your discretion),
	- insn->op_str (at your discretion).

	Note that insn->detail is always set to 0 before calling callback with insn->id == 0.