cs_disasm_single: new function
(1) options SKIPDATA or CHECKINSN are not handled by this function.

(2) this function only handles one instruction per call.

(3) it returns true for a decoded or an invalid instruction (checked with insn->id == 0)

(4) it returns false if something goes wrong with memory.

(5) the first call to cs_disasm_single should be done with parameter insn such as *insn == 0 so the function will preallocate a cs_insn instance with detail if necessary.

(6) the next calls to cs_disasm_single reuses the same preallocated cs_insn instance and the same detail if necessary.

(7) when finished cs_free with count set to 1 must be called.