Add `CMAKE_EXPORT_COMPILE_COMMANDS` to cmake build doc and reverse renaming
This adds the `-DCMAKE_EXPORT_COMPILE_COMMANDS=ON` to the `cmake` build instructions.
I think it is a commonly wanted option since `compile_commands.json` is used by `clangd`. So people who work on the code possibly want this.

Also this renames the file from `debug_build_docs` to `COMPILE_CMAKE.TXT`. Because I accidentally renamed it in the last commit.