capstone.pc.in: re-add capstone include subdirectory via Cflags.
CMake installs the "common headers" for Capstone into a capstone/
subdirectory of the system's includedir:

  install(
    FILES ${HEADERS_COMMON}
    DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/capstone
  )

The pkg-config file, on the other hand, omits the subdirectory:

    includedir=${prefix}/include
    Cflags: -I${includedir}

This means that consumers of the library would need to include the
"capstone/" prefix when including those headers. For example,

  #include <capstone/capstone.h>

rather than

  #include <capstone.h>

Neither is "wrong" -- it's a design choice, but one that breaks the
public API when it changes. This has happened before:

  https://github.com/capstone-engine/capstone/commit/0a39b785
  https://github.com/capstone-engine/capstone/commit/02e33f8b

And (for example) PHP was updated to reflect that, i.e. to work with
capstone-4.x:

  https://github.com/php/php-src/commit/ffc2a53a

But now in the 5.x series of Capstone, the path (or the -I flag,
relative to that path) has reverted,

  https://github.com/capstone-engine/capstone/commit/6656bcb6

breaking PHP once again. This commit re-fixes it by updating the
directory passed via Cflags.

Closes: https://github.com/capstone-engine/capstone/issues/1982