Restore the ARM register naming from v4.
In v5 the register naming for ARM registers changed. This was not a
documented behaviour, and may not be desireable.

This change makes the register names in v5 the same as they were in
v4 - that is that when NOREGNAME is used, R13 and R14 are known as
`sp` and `lr`. They had been changed to `r13` and `r14`.

See https://github.com/capstone-engine/capstone/issues/2078 for more
details.