Add feature to select between alias and real operand set.
Alias instructions always had the operand set of the alias instruction.
This can hide certain important operands.
This commit adds the possibility to print the alias instructions as always,
but have the detail operands set as if it would be the real instructions.

This greatly helps increases the consistent use of detail->operands.