Update and refactor BPF
One of the main improvements recent eBPF verifiers have is the ability to jump backwards in code.
This means that the OFF field of the opcode can be negative, and thus should be interpreted as signed.

eBPF used to have load instructions, specific to loading data from
socket buffer. Today, they are deprecated, but it is still good practice
to support them, as it's very likely older programs might use them.

Some of the newer instructions added by the specification require additional fields to dictate
what is the exact operation to be performed (eg. CALL requires 'src' to decide whether 'imm' should interpreted as BTF function ID,
or as offset from IP. MOV requires 'off' field in order to decide whether a regular MOV or rather a MOVSX should be performed)

* Move BPF opcode setting to the disassembler
* Change BPF JMP instruction naming to conventional JA
* Fix the sign interpretation of 'off', 'imm', etc
* Implement LDIND* and LDABS* legacy packet instructions
* Implement JUMP32 instruction class
* Move BPF opcode setting to disassembler
* Correct and remove branch + load tests, and add more exhaustive ones
* Correct and remove malformed LD tests, and add legacy packet tests