Update read/written registers for x86 procedure return instructions
These should have been part of
https://github.com/capstone-engine/capstone/pull/2578

There are two bugs in 64-bit mode that can't be fixed here:

1. retf (0xcb) is decoded as X86_LRETL instead of X86_LRETQ
2. retf <imm> (0xca iw16) is decoded as X86_LRETIL instead of X86_LRETIQ

In 16-bit mode, Zydis and xed use 'esp' instead of 'sp' for all 'ret*'
instructions. I've opted not to follow that since it makes no sense to
have a 32-bit register present in a 16-bit context. I'll submit a bug
report to Zydis and see what they say.